+++ src/main/java/duke/command/AddDeadlineCommand.java
@@ -1,17 +1,23 @@
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
+package duke.command;
 
-public class AddEventCommand extends Command {
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.Deadline;
+import java.time.LocalDate;
+
+public class AddDeadlineCommand extends Command {
     protected String description;
-    protected String at;
+    protected String by;
     protected LocalDate date;
 
-    public AddEventCommand(String description, String at) {
+    public AddDeadlineCommand(String description, String by) {
         this.description = description;
-        this.at = at;
+        this.by = by;
     }
-    public AddEventCommand(String description, LocalDate date) {
+    public AddDeadlineCommand(String description, LocalDate date) {
         this.description = description;
         this.date = date;
     }
@@ -20,10 +26,10 @@ public class AddEventCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         Task task;
         if (date == null) {
-            task = new Event(description, at);
+            task = new Deadline(description, by);
         }
         else {
-            task = new Event(description, date);
+            task = new Deadline(description, date);
         }
         tasks.addTask(task);
         ui.showAddTask(task, tasks.getTasks());
