+++ src/main/java/duke/command/ListCommand.java
@@ -1,0 +1,12 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class ListCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showList(tasks.getTasks());
+    }
+}
