+++ src/main/java/duke/command/MarkCommand.java
@@ -1,4 +1,9 @@
-import java.io.IOException;
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+import duke.exception.DukeException;
 
 public class MarkCommand extends Command {
     protected int number;
