+++ src/main/java/duke/main/Duke.java
@@ -1,4 +1,11 @@
-import java.io.IOException;
+package duke.main;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+import duke.exception.DukeException;
+import duke.command.Command;
+import duke.parser.Parser;
 
 public class Duke {
