+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,10 @@
+package duke.storage;
+
+import duke.task.Task;
+import duke.exception.DukeException;
+import duke.task.Todo;
+import duke.task.Deadline;
+import duke.task.Event;
 import java.io.*;
 import java.util.ArrayList;
